pyproject.toml
[project]
name = "vention_rail_module"
version = "1.0.0"
description = "Software for automatting a Vention Rails"
authors = [
    {name = "Ryan D. Lewis", email="ryan.lewis@anl.gov"},
    {name = "Rafael Vescovi", email="ravescovi@anl.gov"},
    {name = "Doga Ozgulbas", email="dozgulbas@anl.gov"},
    {name = "Abe Stroka", email="astroka@anl.gov"},
    {name = "Kyle Hippe", email = "khippe@anl.gov"},
    {name = "Tobias Ginsburg", email = "tginsburg@anl.gov"},
]
dependencies = [
    "madsci.node_module>=0.2.1",
    "numpy",
    "paho-mqtt",
]

requires-python = ">=3.9.1"
readme = "README.md"
license = {text = "MIT"}

[project.urls]
homepage = "https://github.com/AD-SDL/vention_rail_module"

[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[dependency-groups]
dev = [
    "ruff>=0.11.7",
    "pre-commit>=4.2.0",
    "pytest>=8.3.5",
]
